draft flask
